Refactor: move out GitLab repo sync code
